fix: user도메인 기능 중복 해결 / 구조 정리#42-user-fix
