+++ src/main/java/devut/buzzerbidder/domain/user/controller/UserMeController.java
@@ -4,9 +4,11 @@ import devut.buzzerbidder.domain.deal.enums.AuctionType;
 import devut.buzzerbidder.domain.deal.service.LiveDealService;
 import devut.buzzerbidder.domain.deliveryTracking.dto.request.DeliveryRequest;
 import devut.buzzerbidder.domain.deliveryTracking.dto.response.DeliveryTrackingResponse;
-import devut.buzzerbidder.domain.user.dto.response.UserInfoResponse;
+import devut.buzzerbidder.domain.user.dto.request.UserUpdateRequest;
+import devut.buzzerbidder.domain.user.dto.response.UserProfileResponse;
+import devut.buzzerbidder.domain.user.dto.response.UserUpdateResponse;
 import devut.buzzerbidder.domain.user.entity.User;
-import devut.buzzerbidder.domain.user.service.MypageService;
+import devut.buzzerbidder.domain.user.service.UserService;
 import devut.buzzerbidder.global.exeption.BusinessException;
 import devut.buzzerbidder.global.exeption.ErrorCode;
 import devut.buzzerbidder.global.requestcontext.RequestContext;
@@ -21,11 +23,11 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1/users/me")
 @RequiredArgsConstructor
 @Tag(name = "MyPage", description = "마이페이지 api")
-public class MyPageController {
+public class UserMeController {
 
     private final RequestContext requestContext;
     private final LiveDealService liveDealService;
-    private final MypageService mypageService;
+    private final UserService userService;
 
     @PatchMapping("/deals/{type}/{dealId}/delivery")
     @Operation(summary = "배송 정보 입력")
@@ -66,12 +68,20 @@ public class MyPageController {
         return trackInfo != null ? ApiResponse.ok("배송조회 성공", trackInfo) : ApiResponse.error(ErrorCode.DEAL_DELIVERY_INFO_NOT_FOUND, null);
     }
 
+    @Operation(summary = "내 정보 조회", description = "현재 로그인한 사용자의 정보를 조회합니다.")
     @GetMapping
-    @Operation(summary = "내 정보 조회")
-    public ApiResponse<UserInfoResponse> getMyInfo() {
+    public ApiResponse<UserProfileResponse> getMyProfile() {
         User currentUser = requestContext.getCurrentUser();
-
-        return ApiResponse.ok("회원정보 조회 성공", mypageService.getUserInfo(currentUser));
+        UserProfileResponse response = userService.getMyProfile(currentUser);
+        return ApiResponse.ok("회원정보 조회 성공", response);
     }
 
+    @Operation(summary = "내 정보 수정", description = "현재 로그인한 사용자의 정보를 수정합니다.")
+    @PatchMapping
+    public ApiResponse<UserUpdateResponse> updateMyProfile(
+            @Valid @RequestBody UserUpdateRequest request) {
+        User currentUser = requestContext.getCurrentUser();
+        UserUpdateResponse response = userService.updateMyProfile(currentUser, request);
+        return ApiResponse.ok("회원정보 수정 성공", response);
+    }
 }
