+++ src/main/java/devut/buzzerbidder/domain/user/dto/request/UserUpdateRequest.java
@@ -15,7 +15,7 @@ public record UserUpdateRequest(
         String nickname,
 
         @Schema(description = "생년월일", example = "2000-01-01")
-        LocalDate birth,
+        LocalDate birthDate,
 
         @Schema(description = "프로필 이미지 URL", example = "https://example.com/image.jpg")
         String image
