+++ src/main/java/devut/buzzerbidder/domain/user/dto/response/UserInfo.java
@@ -20,7 +20,7 @@ public record UserInfo(
         LocalDate birthDate,
 
         @Schema(description = "프로필 이미지 URL", example = "https://example.com/image.jpg")
-        String profileImageUrl
+        String image
 ) {
     public static UserInfo from(User user) {
         return new UserInfo(
